Return number of bytes cid_hex_to_binary decoded
Update cid_hex_to_binary to return the actual number of bytes which were
decoded. The actual number of bytes decoded can be quite variable
depending on the number of ignored 'whitespace' bytes or early
termination with '>'. Update the caller to use this calculated value to
avoid trusting 'parser->binary_length' will always be correct and
reading uninitialized bits if fewer are actually decoded.

First reported as https://crbug.com/1203240 .