Delay setting names and langTags until computed
Previously, the `table->names` and `table->langTags` fields were created
pointing to uninitialized memory and an early exit could happen if the
frame could not be entered. The caller would then be unable to properly
dispose of the memory as the `string` fields had not been initialized.

* src/sfnt/ttload.c (tt_face_load_name): delay setting `table->langTags`
and `table->names` until after the memory they will point to is fully
initialized.