meson: check the return value of run_command
By default, errors are not checked and a command that is somehow broken
will just capture incorrect output (likely an empty string). Current
development versions of meson now raise a warning for this implicit
behavior, and advise explicitly setting the `check: ` kwarg to determine
whether a failing return code should be considered an error.

Since none of the commands in this project are expected to fail, mark
them as required to succeed.