meson: fix name of python executable for auxiliary scripts
The previous change to check the return code of run_command()
invocations caused the CI to fail. Although most scripts used python_exe
as the program command, the script to determine the project version did
not.

But, all scripts used `python` as the shebang, and this is not available
on all systems. Particularly Debian does not provide a `python` command,
though `python3` does exist. This meant that formerly the version number
was lacking, and now the build simply fails.

Instead, rely on `python3` since:
- it is guaranteed to exist when running meson
- python2 is end of life anyway