Meson: Remove default_library=both from default_options
It is inconsistent with what all other Meson projects does.

It also prevents static linking to freetype when it is a subproject.
That's because default_options in a subproject takes precedence on
values set on the main project. For example building GStreamer with
`-Ddefault_library=static` would still dynamic link on freetype, unless
user also set `-Dfreetype2:default_library=static`.