[sfnt, truetype] Don't require gvar for font to support variations
Fixes #1172.

* src/sfnt/sfobjs.c (sfnt_load_face): Tag font as multiple masters font
just if `fvar` is present, do not require other tables to be resent.
* src/truetype/ttgxvar.c (tt_set_mm_blend): Allow for a missing `gvar`
table when setting variation coordinates but if a `gvar` table is
present still sanity check it and fail on malformedness.
(TT_Get_MM_Var): Don't assume `fvar` needs `gvar` or `CFF2` in all
cases, which is an overly tight check.