[truetype] Reset cvt and storage in context load
Currently the cvt and storage are saved and restored in `TT_RunIns`.
However, this is too granular as the cvt and storage should be set to
the original cvt and storage only when setting up the hinting context.
This allows for the cvt and storage to be modified while parsing
multiple glyphs, as is the case with composite glyphs.

* src/truetype/ttinterp.h (TT_ExecContextRec_): remove `origCvt` and
`origStorage`

* src/truetype/ttinterp.c (TT_RunIns): don't save and restore the cvt
and storage. (Modify_CVT_Check, Ins_WS): switch from "if in glyph and
using original data do copy on write" to "if in glyph and not using
glyph specific data do copy on write".