[base] Error when requested driver not found.
In `open_face_from_buffer` it is possible that a driver is requested but
FreeType was built without the requested module. Return an error in this
case to indicate that the request could not be satisfied, rather than
trying all existing driver modules.

* src/base/ftobjs.c (open_face_from_buffer): return `Missing_Module` if
a driver is specified but not found.